Moved entities to own package
